refactor search-related stuff
